Сущности в отдельный файл Реализована миграция
