+++ LibraryEFCore/AppContext.cs
@@ -7,34 +7,17 @@ using System.Threading.Tasks;
 
 namespace LibraryEFCore
 {
-    // класс-модель Авторы
-    public class Authors
-    {
-        public int ID { get; set; }
-        public string FullName { get; set; }
-        public int YearOfBirth { get; set; }
-        public List<Books> Books { get; set; } = new List<Books>(); // Связь Many-to-many 
-    }
-    // класс-модель Книги
-    public class Books
-    {
-        public int ID { get; set; }
-        public string Name { get; set; }
-        public int PublicationYear { get; set; }
-        public List<Authors> Authors { get; set; } = new List<Authors>(); // Связь Many-to-many 
-    }
-
     public class AppContext: DbContext
     {
         // Объявляем наборы объектов, через которые идет связь с соответствующей по имени таблицей в БД
         public DbSet<Authors> Authors { get; set; }
         public DbSet<Books> Books { get; set; }
 
-        // Конструктор, где пересоздаем БД
+        // Конструктор, где пересоздаем БД. При миграции не используем 
         public AppContext()
         {
-            Database.EnsureDeleted();
-            Database.EnsureCreated();
+            //Database.EnsureDeleted();
+            //Database.EnsureCreated();
         }       
 
         // Переопределяем метод для передачи строки подключения
